delete pickupPoint done

+++ gateway/gateway/Logic/Implementation/LocationLogicImpl.cs
@@ -1,3 +1,4 @@
+using CSharpShared.Exception;
 using gateway.DTO;
 using gateway.RpcClient.Interface;
 
@@ -6,13 +7,15 @@ namespace gateway.Model.Implementation;
 public class LocationLogicImpl : ILocationServiceLogic
 {
     private readonly ILocationServiceClient _locationServiceClient;
+    private readonly IPackageServiceClient _packageServiceClient;
     private readonly DtoMapper _dtoMapper;
     private readonly Logger.Logger _logger = Logger.Logger.Instance;
 
-    public LocationLogicImpl(ILocationServiceClient locationServiceClient, DtoMapper dtoMapper)
+    public LocationLogicImpl(ILocationServiceClient locationServiceClient, DtoMapper dtoMapper, IPackageServiceClient packageServiceClient)
     {
         _locationServiceClient = locationServiceClient;
         _dtoMapper = dtoMapper;
+        _packageServiceClient = packageServiceClient;
     }
 
     public async Task<IEnumerable<GetPickUpPointDto>> GetAllPickUpPointsAsync()
@@ -25,7 +28,17 @@ public class LocationLogicImpl : ILocationServiceLogic
 
     public async Task DeletePickupPoint(long id)
     {
-        await _locationServiceClient.DeleteLocation(id);
+        bool locationCanNotBeEdited = false;
+        var packagesUsingThisLocation = await _packageServiceClient.GetPackagesByLocationIdAsync(id);
+        if (packagesUsingThisLocation.Packet.Count == 0)
+        {
+            await _locationServiceClient.DeleteLocation(id);
+            
+        }
+        else
+        {
+            throw new LocationUsedException();
+        }
         _logger.Log($"LocationLogicImpl: DeleteLocation was called with id:  {id}");
     }
 }
